requirements.txt
mediapipe
opencv-python
pandas
scikit-learn
google-generativeai
numpy
